+++ Gateway/Validator/RefundCodeValidator.php
@@ -1,0 +1,49 @@
+<?php
+
+namespace StancerIntegration\Payments\Gateway\Validator;
+
+use Magento\Payment\Gateway\Validator\AbstractValidator;
+use Magento\Payment\Gateway\Validator\ResultInterface;
+use Magento\Payment\Gateway\Validator\ResultInterfaceFactory;
+use Stancer\Payment;
+use StancerIntegration\Payments\Gateway\Helper\SubjectReader;
+
+class RefundCodeValidator extends AbstractValidator
+{
+    /**
+     * @var SubjectReader
+     */
+    private $subjectReader;
+
+    /**
+     * @param ResultInterfaceFactory $resultFactory
+     * @param SubjectReader $subjectReader
+     */
+    public function __construct(
+        ResultInterfaceFactory $resultFactory,
+        SubjectReader          $subjectReader
+    ) {
+        parent::__construct($resultFactory);
+        $this->subjectReader = $subjectReader;
+    }
+
+    /**
+     * Performs validation of result code
+     *
+     * @param array $validationSubject
+     * @return ResultInterface
+     */
+    public function validate(array $validationSubject)
+    {
+        /** @var Payment $response */
+        $response = $this->subjectReader->readRefund($validationSubject['response']);
+        if (!empty($response->getId())) {
+            return $this->createResult(true, []);
+        } else {
+            return $this->createResult(
+                false,
+                [__('Gateway rejected the transaction.')]
+            );
+        }
+    }
+}